First commit to initialize all file into repository
Create a new rails project call agendaedu-message-box and initialize
this skeleton of ruby on rails project agendaedu-message-box as a git
repository with  all standards files and structure of
ruby on rails project.

To do that it was execute follow line command:
docker run -it --rm --user "$(id -u):$(id -g)" -v "$PWD":/usr/src/app
  -w /usr/src/app rails rails new --skip-bundle agendaedu-message-box
  --database=postgresql

As a result, we have all initialize this new structure ruby on rails
project into a git repository.