Configure database
Configure database to be ready to create and migrate some data. To do
that the file config/database.yml was modify to include configuration
about database in development, test and production. The only important
modification was setup host and user variable to postgres, this is a
link to docker service to postgres.

As a result, this application it's ready to create database to be
consume for all operations that this application it will need to
be construct.